% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_locs}
\alias{clean_locs}
\title{Clean PESC site names}
\usage{
clean_locs(x)
}
\arguments{
\item{x}{A string of PESC site names.}
}
\value{
A cleaned string containing BPFA site names.
}
\description{
Site names are slightly different and formatted differently
between the sample sheets PESC sends vs the site names in
the BPFA location table. This function quickly cleans and
standardizes PESC site names to match BPFA site names.
}
