% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_samples}
\alias{clean_samples}
\title{Clean PESC sample names}
\usage{
clean_samples(x)
}
\arguments{
\item{x}{A string of PESC sample names.}
}
\value{
A dataframe consisting of the original sample name, sample bag number (if detected), waypoint, replicate status (if detected), and two logical columns indicating if the sample was collected in Jensen's Bay (`jb == TRUE`) or Maltby Slough (`mb == TRUE`).
}
\description{
PESC sample names do not match BPFA sample names and need to be
cleaned prior to being matched to BPFA sample metadata. This
function extracts out the waypoint number, bag number (if present),
replicate (if present), and cleans up any superfluous text/characters.
}
\details{
Some samples are run multiple times, e.g. have are replicated.
In these cases PESC will append the word "dup" to the end of the
sample name to indicate it is a duplicate. If duplicate samples
are detected, the function will note that the sample is a replicate.

PESC often reports sample locations as "TOFINO"; however, BPFA
locations differentiate between Jensen's Bay and Maltby Slough.
PESC deals with this by appending "JB" or "MS" to the start of
sample names. If the function detects if Jensen's Bay or Maltby Slough
location information is contained within the sample name string,
the function returns this information in the 'jb' and 'ms' column.
}
