% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_bpfa}
\alias{connect_bpfa}
\title{Connect to the BPFA database}
\usage{
connect_bpfa()
}
\value{
An SQLiteConnection object.
}
\description{
A simple wrapper for `DBI::dbConnect(RSQLite::SQLite(), file.path(bpfa_dir(), "bpfa.db"))`.
}
\examples{
\dontrun{
bpfa <- connect_bpfa()
DBI::dbListTables(bpfa)
DBI::dbDisconnect(bpfa)
}
}
