% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{copy_bpfa}
\alias{copy_bpfa}
\title{Save a copy of the BPFA database to your local machine}
\usage{
copy_bpfa(path)
}
\arguments{
\item{path}{A file path to the copy of the database you wish to save.}
}
\description{
This function will update your local BPFA database with a new copy of the database.
For example, if you were emailed a copy of the database with updated data, you would
use `copy_bpfa()` to update your local database with the copy you received via email.
Your local BPFA database (the one that the R package `{bpfa}` uses) will then be updated
and usable with any `{bpfa}` package functions.
}
\examples{
\dontrun{
copy_bpfa("downloads/new_bpfa.db")}
}
