% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{import_bpfa}
\alias{import_bpfa}
\title{Import processed PESC data into the BPFA database}
\usage{
import_bpfa(lims_out, verbose = TRUE)
}
\arguments{
\item{lims_out}{Output from the read_lims function.}

\item{verbose}{Logical (T/F). Should the number of new records appended into each table be printed?}
}
\description{
After data are read in and cleaned by the `read_lims` function,
they can then be added to the three PESC data tables in the BPFA
database: `pesc_batch`, `pesc_benchtop`, and `pesc_data`.
}
\examples{
\dontrun{
initialize_bpfa()
lims <- read_lims('~Documents/path/to/lims/file.xlsx')
import_bpfa(lims)
}
}
