% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{initialize_bpfa}
\alias{initialize_bpfa}
\title{Create a new BPFA database}
\usage{
initialize_bpfa()
}
\description{
This function will create a new, clean copy of the Brunswick Point
Fatty Acid (BPFA) database. A new copy of the database comes with
the location and sample tables pre-loaded, but no PESC results data.
}
\details{
If the function detects that you already have a copy of the BPFA
database on your computer, it will ask if you would like to keep
a backup copy of the existing database or if you would like to
overwrite it instead.
}
\examples{
\dontrun{
initialize_bpfa()
}
}
