% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge full sample metadata with PESC data}
\usage{
merge_samples(lims_out, samples)
}
\arguments{
\item{lims_out}{List output from read_lims()}

\item{samples}{Fatty acid sample metadata following the BPFA data entry template}
}
\value{
A list of containing PESC data matched to full sample metadata, PESC results, and PESC bench sheets containing test tube metadata.
}
\description{
Merge full sample metadata with PESC data
}
\examples{
\dontrun {
out <- read_lims("~/Documents/path/to/lims/file.xslx")
samples <- read.csv("~/Documents/path/to/sample/metadata.csv")
merged_results <- merge_samples(lims_out = out, samples = samples)
}
}
