% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{update_samples}
\alias{update_samples}
\alias{update_locations}
\alias{update_metadata}
\title{Update BPFA sample metadata}
\usage{
update_samples()

update_locations()

update_metadata()
}
\value{

}
\description{
Sometimes the metadata for samples, e.g. locations or sampling dates,
needs to be updated in the database. If the metadata bundled with the package
has been updated in a new package release, these functions can be used to
quickly write those updates to your local copy of the BPFA database.

Note that sample locations are stored in a seperate table from the sample
information itself, because multiple samples are collected repeatedly from the
sample field sites and waypoints.
}
\section{Functions}{
\itemize{
\item \code{update_samples}: Update sample information (e.g., sampling date, sampler, weather conditions)

\item \code{update_locations}: Update sampling locations (lat/long)

\item \code{update_metadata}: Update both sample information and sample locations
}}

