% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{bpfa_dir}
\alias{bpfa_dir}
\title{Output path to the BPFA sqlite database}
\usage{
bpfa_dir(...)
}
\arguments{
\item{...}{arguments potentially passed to \code{rappdirs::user_data_dir}.}
}
\description{
Prints the BPFA sqlite file storage location, regardless of OS, for \link{initialize_bpfa}.
}
\examples{
\dontrun{
bpfa_dir()
}
}
