% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_lims}
\alias{read_lims}
\title{Read and clean up LIMS data}
\usage{
read_lims(path)
}
\arguments{
\item{path}{Path to an Excel file containing LIMS data.}
}
\value{
A list containing two dataframes.
}
\description{
Read and clean up LIMS data
}
\examples{
\dontrun{
read_lims("~/Documents/path/to/lims/file.xslx")
}
}
